requirements.txt
streamlit
pdfminer.six
PyPDF2
biopython
requests
pandas
